fix bug

+++ clean_git_merged_branch.sh
@@ -7,9 +7,11 @@ source log.sh
 
 DEFAULT_commit=HEAD
 DEFAULT_include_remotes=origin
+DEFAULT_exclude_branches=master
 
 
 source getopt.sh
+
 
 #set -o verbose
 #set -e
@@ -22,9 +24,11 @@ declare -a remote_branches=(`echo "$branches" | grep    "^remotes\/" | cut -d\/ 
 declare branch
 declare idx
 declare out_lines
+declare _c_exclude_branches_lines=`echo "${exclude_branches}" | tr ' ' '\n'`
 for idx in ${!local_branches[@]}
 do
     branch=${local_branches[$idx]}
+    if echo "$_c_exclude_branches_lines" | grep -sqFx "$branch"; then continue; fi
     log_info "Clean local merged branch \"$branch\"." && {
         out_lines=`git branch -d $branch 2>&1`
         test $? -eq 0 || \
@@ -43,10 +47,9 @@ for idx in ${!remote_branches[@]}
 do
     item=${remote_branches[$idx]}
     remote=`echo $item | cut -d\/ -f1` && \
-    if [ -n "$include_remotes" ]; then
-        echo "${_c_include_remotes_lines}" | grep -sqFx "$remote" || continue;
-    fi && \
+    if [ -n "$include_remotes" ] && ! echo "${_c_include_remotes_lines}" | grep -sqFx "$remote"; then continue; fi
     branch=`echo $item | cut -d\/ -f2` && \
+    if echo "$_c_exclude_branches_lines" | grep -sqFx "$branch"; then continue; fi
     log_info "Clean merged remote branch \"$item\"." && {
         out_lines=`git push --delete $remote $branch 2>&1`
         test $? -eq 0 || \
@@ -60,6 +63,3 @@ do
     git branch -r -d $item && \
     unset remote_branches[$idx]
 done
-
-declare -p local_branches
-declare -p remote_branches
